move sio outside main
